add comments and functions descriptions to utils.py and change file name to data_store.py
